New function - 'read_batterylevel'
The read_batterylevel function reads the ADC_BATTERY pin on the MRK1010
board and converts the analog value 0-1023 to a voltage from the bus,
with 3.3v as reference.

The percentage is calculated by estimating the max current of a standard
3.7v LiPo battery, approx. 3.9 - 4.2 volts, and calculating the actual
current on the bus by this reference value to estimate remaining
charge.